+++ components/zcash_protocol/src/consensus.rs
@@ -7,7 +7,7 @@ use std::fmt;
 use std::ops::{Add, Bound, RangeBounds, Sub};
 use zcash_address;
 
-use crate::{constants, sapling::note_encryption::Zip212Enforcement};
+use crate::constants;
 
 /// A wrapper type representing blockchain heights.
 ///
@@ -632,25 +632,6 @@ impl BranchId {
     }
 }
 
-/// Returns the enforcement policy for ZIP 212 at the given height.
-pub fn sapling_zip212_enforcement(
-    params: &impl Parameters,
-    height: BlockHeight,
-) -> Zip212Enforcement {
-    if params.is_nu_active(NetworkUpgrade::Canopy, height) {
-        let grace_period_end_height =
-            params.activation_height(NetworkUpgrade::Canopy).unwrap() + ZIP212_GRACE_PERIOD;
-
-        if height < grace_period_end_height {
-            Zip212Enforcement::GracePeriod
-        } else {
-            Zip212Enforcement::On
-        }
-    } else {
-        Zip212Enforcement::Off
-    }
-}
-
 #[cfg(any(test, feature = "test-dependencies"))]
 pub mod testing {
     use proptest::sample::select;
@@ -688,6 +669,7 @@ pub mod testing {
             })
     }
 
+    #[cfg(feature = "test-dependencies")]
     impl incrementalmerkletree::testing::TestCheckpoint for BlockHeight {
         fn from_u64(value: u64) -> Self {
             BlockHeight(u32::try_from(value).expect("Test checkpoint ids do not exceed 32 bits"))